refactor && feat: refactor my info page of user dashboard's structure and also connect profile photo section to backend
